+++ tests/benchmark/llm_utils/__init__.py
@@ -7,6 +7,6 @@ from llm_utils.decode_utils import (
     extract_topk,
     generate_and_benchmark,
     init_accuracy_testing,
+    init_mla_cache,
     init_static_cache,
 )
-from llm_utils.mla_utils import MLACache, MLAStaticLayer, init_mla_cache
