+++ tests/torch/models/kimi_k2/test_kimi_k2.py
@@ -3,6 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
+import sys as _sys
+from pathlib import Path as _Path
 
 import numpy as np
 import pytest
@@ -17,7 +19,6 @@ from transformers import DynamicCache
 from tt_torch.sparse_mlp import create_a2a_from_deepseek_v3_moe, enable_sparse_mlp
 
 from tests.utils import failed_ttmlir_compilation
-from tt_torch.sparse_mlp import create_a2a_from_deepseek_v3_moe
 
 from .configuration_deepseek import DeepseekV3Config
 from .modeling_deepseek import (
@@ -28,7 +29,9 @@ from .modeling_deepseek import (
     DeepseekV3MoE,
 )
 from .original_modeling_deepseek import DeepseekV3Attention as OrigDeepseekV3Attention
-from .utils import MLACache
+
+_sys.path.insert(0, str(_Path(__file__).resolve().parents[2] / "utils"))
+from mla_cache import MLACache
 
 
 @pytest.mark.xfail(
@@ -555,306 +558,3 @@ def test_kimi_k2_mla_cache():
 
     assert torch.equal(mla_key, orig_key)
     assert torch.equal(mla_val, orig_val)
-
-
-@pytest.mark.nightly
-def test_kimi_k2_causal_lm_mla_cache_prefill_decode():
-    """
-    CPU test validating prefill + decode of DeepseekV3ForCausalLM with MLACache.
-
-    Uses a 2-layer model (layer 0 = dense MLP, layer 1 = MoE) with a small
-    config to keep the test fast. Verifies that:
-      - Prefill fills the cache and returns logits of the right shape.
-      - Multiple decode steps advance cache_position and return single-token logits.
-      - The cache is actually written to (non-zero entries after prefill).
-    """
-    config = DeepseekV3Config(
-        vocab_size=100,
-        hidden_size=64,
-        intermediate_size=128,
-        moe_intermediate_size=32,
-        num_hidden_layers=2,
-        num_attention_heads=4,
-        num_key_value_heads=4,
-        q_lora_rank=32,
-        kv_lora_rank=16,
-        qk_rope_head_dim=16,
-        v_head_dim=8,
-        qk_nope_head_dim=8,
-        n_shared_experts=1,
-        n_routed_experts=4,        # 4 experts total
-        num_experts_per_tok=2,     # top-2 selected per token
-        first_k_dense_replace=1,   # layer 0 dense, layer 1 MoE
-        n_group=2,                 # 2 groups of 2 experts each (noaux_tc requires >=2 per group)
-        topk_group=1,              # pick from 1 group
-        max_position_embeddings=64,
-        rope_scaling=None,
-        attention_bias=False,
-        attention_dropout=0.0,
-    )
-
-    BATCH_SIZE = 2
-    PREFILL_LEN = 8
-    DECODE_STEPS = 3
-    MAX_CACHE_LEN = PREFILL_LEN + DECODE_STEPS
-
-    model = DeepseekV3ForCausalLM(config).eval()
-
-    mla_cache = MLACache(config, max_cache_len=MAX_CACHE_LEN)
-
-    # --- Prefill ---
-    prefill_tokens = torch.randint(0, config.vocab_size, (BATCH_SIZE, PREFILL_LEN))
-    prefill_cache_position = torch.arange(PREFILL_LEN)
-
-    with torch.no_grad():
-        prefill_output = model(
-            input_ids=prefill_tokens,
-            past_key_values=mla_cache,
-            cache_position=prefill_cache_position,
-            use_cache=True,
-        )
-
-    assert prefill_output.logits.shape == (BATCH_SIZE, PREFILL_LEN, config.vocab_size)
-
-    # Cache must be initialized and written to after prefill
-    for layer_cache in mla_cache.layers:
-        assert layer_cache.is_initialized
-        assert layer_cache.compressed_kv[:, 0, :PREFILL_LEN, :].abs().sum() > 0
-
-    # --- Decode ---
-    next_token = prefill_output.logits[:, -1].argmax(dim=-1, keepdim=True)
-
-    for step in range(DECODE_STEPS):
-        decode_cache_position = torch.tensor([PREFILL_LEN + step])
-
-        with torch.no_grad():
-            decode_output = model(
-                input_ids=next_token,
-                past_key_values=mla_cache,
-                cache_position=decode_cache_position,
-                use_cache=True,
-            )
-
-        assert decode_output.logits.shape == (BATCH_SIZE, 1, config.vocab_size)
-        next_token = decode_output.logits[:, -1].argmax(dim=-1, keepdim=True)
-
-
-@pytest.mark.nightly
-@pytest.mark.galaxy
-def test_kimi_k2_moe_compile():
-    """
-    Torch graph test for DeepseekV3MoE via SparseMLP compiled and run on Galaxy (4x8 mesh).
-
-    Uses a small config (hidden_size=64, 4 routed experts) to keep instantiation fast.
-    The raw DeepseekV3MoE cannot be compiled because moe_infer uses
-    tokens_per_expert.cpu().tolist() for data-dependent loop control, which breaks
-    dynamo's fake-tensor tracing. This test converts the MoE to A2aSparseMLPWithSharedExperts
-    via create_a2a_from_deepseek_v3_moe, which replaces the per-expert Python loop with
-    torch.ops.tt.sparse_matmul and static tensor routing.
-    """
-    xr.set_device_type("TT")
-    torch_xla.runtime.use_spmd()
-    os.environ["CONVERT_SHLO_TO_SHARDY"] = "1"
-
-    config = DeepseekV3Config(
-        hidden_size=64,
-        intermediate_size=128,
-        moe_intermediate_size=32,
-        num_hidden_layers=2,
-        num_attention_heads=4,
-        num_key_value_heads=4,
-        q_lora_rank=32,
-        kv_lora_rank=16,
-        qk_rope_head_dim=16,
-        v_head_dim=8,
-        qk_nope_head_dim=8,
-        n_shared_experts=1,
-        n_routed_experts=32,
-        num_experts_per_tok=2,
-        first_k_dense_replace=1,
-        n_group=4,
-        topk_group=1,
-        max_position_embeddings=64,
-        rope_scaling=None,
-        attention_bias=False,
-        attention_dropout=0.0,
-    )
-
-    moe = DeepseekV3MoE(config).to(torch.bfloat16).eval()
-
-    BATCH_SIZE = 32
-    SEQ_LEN = 16
-
-    num_devices = xr.global_runtime_device_count()
-    mesh_shape = (4, 8)
-    device_ids = np.array(range(num_devices))
-    mesh = Mesh(device_ids, mesh_shape, ("batch", "model"))
-
-    # Convert to A2aSparseMLP — replaces the data-dependent moe_infer loop with
-    # sparse_matmul. cluster_axis=0 dispatches along the batch axis (4 devices),
-    # dispatch_devices=mesh_shape[0]=4.
-    sparse_moe = create_a2a_from_deepseek_v3_moe(
-        moe_module=moe,
-        config=config,
-        num_devices=num_devices,
-        cluster_axis=0,
-        dispatch_devices=mesh_shape[0],
-    ).eval()
-
-    hidden_states = torch.randn(
-        BATCH_SIZE, SEQ_LEN, config.hidden_size, dtype=torch.bfloat16
-    )
-
-    def get_shard_spec(sparse_moe, args, kwargs):
-        shard_specs = {}
-        shard_specs[args[0]] = ("batch", None, None)
-        # Gate (router) weight — replicated across model axis
-        shard_specs[sparse_moe.mlp.router.gate.weight] = (None, "batch")
-        # Stacked expert weights — compound-sharded on expert dim across both axes
-        shard_specs[sparse_moe.mlp.experts.gate_up_proj] = (("batch", "model"), None, None)
-        shard_specs[sparse_moe.mlp.experts.gate_up_proj_bias] = (("batch", "model"), None)
-        shard_specs[sparse_moe.mlp.experts.down_proj] = (("batch", "model"), None, None)
-        shard_specs[sparse_moe.mlp.experts.down_proj_bias] = (("batch", "model"), None)
-        # Shared experts — column/row parallel
-        if sparse_moe.shared_experts is not None:
-            shard_specs[sparse_moe.shared_experts.gate_proj.weight] = ("model", "batch")
-            shard_specs[sparse_moe.shared_experts.up_proj.weight] = ("model", "batch")
-            shard_specs[sparse_moe.shared_experts.down_proj.weight] = ("batch", "model")
-        return shard_specs
-
-    run_graph_test(
-        sparse_moe,
-        [hidden_states],
-        framework=Framework.TORCH,
-        mesh=mesh,
-        shard_spec_fn=get_shard_spec,
-    )
-
-
-def _kimi_k2_shard_model(model, mesh):
-    """Apply tensor-parallel sharding to a Kimi K2 model using the given mesh.
-
-    Mesh axis names must be ("batch", "model").
-    """
-    xs.mark_sharding(model.model.embed_tokens.weight, mesh, (None, "batch"))
-    xs.mark_sharding(model.model.norm.weight, mesh, ("batch",))
-    xs.mark_sharding(model.lm_head.weight, mesh, ("model", "batch"))
-
-    for layer in model.model.layers:
-        # MLA attention
-        xs.mark_sharding(layer.self_attn.q_a_proj.weight, mesh, (None, "batch"))
-        xs.mark_sharding(layer.self_attn.q_b_proj.weight, mesh, ("model", None))
-        xs.mark_sharding(layer.self_attn.q_a_layernorm.weight, mesh, (None,))
-        xs.mark_sharding(layer.self_attn.kv_a_proj_with_mqa.weight, mesh, (None, "batch"))
-        xs.mark_sharding(layer.self_attn.kv_b_proj.weight, mesh, ("model", None))
-        xs.mark_sharding(layer.self_attn.kv_a_layernorm.weight, mesh, (None,))
-        xs.mark_sharding(layer.self_attn.o_proj.weight, mesh, ("batch", "model"))
-
-        if isinstance(layer.mlp, DeepseekV3MoE):
-            xs.mark_sharding(layer.mlp.gate.weight, mesh, (None, "batch"))
-            for expert in layer.mlp.experts:
-                xs.mark_sharding(expert.gate_proj.weight, mesh, ("model", "batch"))
-                xs.mark_sharding(expert.up_proj.weight, mesh, ("model", "batch"))
-                xs.mark_sharding(expert.down_proj.weight, mesh, ("batch", "model"))
-            if getattr(layer.mlp, "shared_experts", None) is not None:
-                xs.mark_sharding(layer.mlp.shared_experts.gate_proj.weight, mesh, ("model", "batch"))
-                xs.mark_sharding(layer.mlp.shared_experts.up_proj.weight, mesh, ("model", "batch"))
-                xs.mark_sharding(layer.mlp.shared_experts.down_proj.weight, mesh, ("batch", "model"))
-        else:
-            xs.mark_sharding(layer.mlp.gate_proj.weight, mesh, ("model", "batch"))
-            xs.mark_sharding(layer.mlp.up_proj.weight, mesh, ("model", "batch"))
-            xs.mark_sharding(layer.mlp.down_proj.weight, mesh, ("batch", "model"))
-
-        xs.mark_sharding(layer.input_layernorm.weight, mesh, ("batch",))
-        xs.mark_sharding(layer.post_attention_layernorm.weight, mesh, ("batch",))
-
-
-@pytest.mark.nightly
-@pytest.mark.galaxy
-def test_kimi_k2_causal_lm_mla_cache_compiled_galaxy():
-    """
-    Compiled prefill + decode for a 2-layer Kimi K2 model with MLACache on Galaxy (4x8 mesh).
-
-    Uses the real Kimi K2 config (hidden_size=7168, 384 experts, …) with
-    num_hidden_layers=2 so we can see which XLA graphs fail to compile.
-    Runs prefill (batch=32, seq=16) then three decode steps.
-    """
-    xr.set_device_type("TT")
-    xr.use_spmd()
-    os.environ["CONVERT_SHLO_TO_SHARDY"] = "1"
-
-    config_path = os.path.join(os.path.dirname(__file__), "config.json")
-    config = DeepseekV3Config.from_json_file(config_path)
-    config.num_hidden_layers = 2
-
-    BATCH_SIZE = 32
-    PREFILL_LEN = 16
-    DECODE_STEPS = 3
-    MAX_CACHE_LEN = PREFILL_LEN + DECODE_STEPS
-
-    # Galaxy mesh (4 rows × 8 columns)
-    num_devices = xr.global_runtime_device_count()
-    mesh_shape = (4, 8)
-    device_ids = np.array(range(num_devices))
-    mesh = Mesh(device_ids, mesh_shape, ("batch", "model"))
-
-    device = torch_xla.device()
-
-    # Build model in bfloat16
-    model = DeepseekV3ForCausalLM(config).to(torch.bfloat16).eval()
-    model = model.to(device)
-    _kimi_k2_shard_model(model, mesh)
-
-    # Pre-initialise MLA cache on CPU then transfer to device
-    mla_cache = MLACache(config, max_cache_len=MAX_CACHE_LEN)
-    dummy_kv = torch.zeros((BATCH_SIZE, 1, 1, config.kv_lora_rank), dtype=torch.bfloat16)
-    dummy_pe = torch.zeros((BATCH_SIZE, 1, 1, config.qk_rope_head_dim), dtype=torch.bfloat16)
-    for cache_layer in mla_cache.layers:
-        cache_layer.lazy_initialization(dummy_kv, dummy_pe)
-
-    for cache_layer in mla_cache.layers:
-        cache_layer.compressed_kv = cache_layer.compressed_kv.to(device)
-        cache_layer.k_pe = cache_layer.k_pe.to(device)
-        cache_layer.keys = cache_layer.compressed_kv
-        cache_layer.values = cache_layer.k_pe
-        torch._dynamo.mark_static_address(cache_layer.compressed_kv)
-        torch._dynamo.mark_static_address(cache_layer.k_pe)
-        # compressed_kv / k_pe are in the shared latent space (not head-sharded);
-        # replicate across the model axis and shard across the batch axis.
-        xs.mark_sharding(cache_layer.compressed_kv, mesh, (None, None, None, None))
-        xs.mark_sharding(cache_layer.k_pe, mesh, (None, None, None, None))
-
-    compiled_model = torch.compile(model, backend="tt")
-
-    # --- Prefill ---
-    prefill_tokens = torch.randint(0, config.vocab_size, (BATCH_SIZE, PREFILL_LEN)).to(device)
-    prefill_cache_position = torch.arange(PREFILL_LEN).to(device)
-
-    with torch.no_grad():
-        prefill_output = compiled_model(
-            input_ids=prefill_tokens,
-            past_key_values=mla_cache,
-            cache_position=prefill_cache_position,
-            use_cache=True,
-        )
-
-    prefill_logits = prefill_output.logits.to("cpu")
-    assert prefill_logits.shape == (BATCH_SIZE, PREFILL_LEN, config.vocab_size)
-
-    # --- Decode ---
-    next_token = prefill_logits[:, -1].argmax(dim=-1, keepdim=True).to(device)
-
-    for step in range(DECODE_STEPS):
-        decode_cache_position = torch.tensor([PREFILL_LEN + step]).to(device)
-
-        with torch.no_grad():
-            decode_output = compiled_model(
-                input_ids=next_token,
-                past_key_values=mla_cache,
-                cache_position=decode_cache_position,
-                use_cache=True,
-            )
-
-        decode_logits = decode_output.logits.to("cpu")
-        assert decode_logits.shape == (BATCH_SIZE, 1, config.vocab_size)
-        next_token = decode_logits[:, -1].argmax(dim=-1, keepdim=True).to(device)
